# backend/api/src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"WordPairInput": {
				"description": "This file defines the types used for deck and wordpair interactions.\n\nNaming Conventions:\n-------------------\nDeck Types:\n- DeckCreateInput: Payload for deck insertion. Contains basic fields and an optional list of wordpairs.\n- DeckUpdateInput: Payload for updating a deck. Contains basic fields (without timestamps) and an optional wordpairs list.\n- DeckEntity:   Represents a deck as stored in the database.\n- DeckSummary:  Extends DeckEntity with the computed \"wordpairCount\". \n- DeckDetail:   Extends DeckSummary by including the full list of associated wordpairs.\n- DeckOptionalReturn: Similar to DeckDetail but with \"wordpairCount\" and \"wordpairs\" optional.\n\nWordPair Types:\n- WordPairInput:  Minimal data required for inserting/updating a wordpair.\n- WordPairSummary: A lightweight version of a wordpair (includes id, deckId, and the word values).\n- WordPairEntity:  Full representation of a wordpair including timestamps.",
				"properties": {
					"wordOriginal": {
						"type": "string"
					},
					"wordTranslation": {
						"type": "string"
					},
					"position": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"wordOriginal",
					"wordTranslation"
				],
				"type": "object",
				"additionalProperties": true
			},
			"LLMDeck": {
				"properties": {
					"name": {
						"type": "string"
					},
					"languageFrom": {
						"type": "string"
					},
					"languageTo": {
						"type": "string"
					},
					"wordpairs": {
						"items": {
							"$ref": "#/components/schemas/WordPairInput"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"languageFrom",
					"languageTo",
					"wordpairs"
				],
				"type": "object",
				"additionalProperties": true
			},
			"apiSuccessResponse_LLMDeck_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/LLMDeck"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"GenerateDeckRequest": {
				"properties": {
					"languageFrom": {
						"type": "string"
					},
					"languageTo": {
						"type": "string"
					},
					"pairCount": {
						"type": "number",
						"format": "double"
					},
					"theme": {
						"type": "string"
					},
					"additionalPrompt": {
						"type": "string"
					}
				},
				"required": [
					"languageFrom",
					"languageTo",
					"pairCount",
					"theme",
					"additionalPrompt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"RefineDeckRequest": {
				"properties": {
					"prompt": {
						"type": "string"
					},
					"history": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"preserveExistingPairs": {
						"type": "boolean"
					},
					"currentDeck": {
						"$ref": "#/components/schemas/LLMDeck"
					}
				},
				"required": [
					"prompt",
					"history",
					"preserveExistingPairs",
					"currentDeck"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ExtractNameResponse": {
				"properties": {
					"name": {
						"type": "string"
					},
					"languageFrom": {
						"type": "string"
					},
					"languageTo": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"languageFrom",
					"languageTo"
				],
				"type": "object",
				"additionalProperties": true
			},
			"apiSuccessResponse_ExtractNameResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/ExtractNameResponse"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ExtractNameRequest": {
				"properties": {
					"wordpairs": {
						"items": {
							"$ref": "#/components/schemas/WordPairInput"
						},
						"type": "array"
					}
				},
				"required": [
					"wordpairs"
				],
				"type": "object",
				"additionalProperties": true
			},
			"DeckSummary": {
				"description": "DeckSummary extends DeckEntity with the computed wordpairCount.\nThis type is ideal for list endpoints where a count is necessary.",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"languageFrom": {
						"type": "string"
					},
					"languageTo": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"lastModified": {
						"type": "string"
					},
					"wordpairCount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"languageFrom",
					"languageTo",
					"createdAt",
					"lastModified",
					"wordpairCount"
				],
				"type": "object",
				"additionalProperties": true
			},
			"apiSuccessResponse_DeckSummary-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/DeckSummary"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"WordPairEntity": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"deckId": {
						"type": "number",
						"format": "double"
					},
					"wordOriginal": {
						"type": "string"
					},
					"wordTranslation": {
						"type": "string"
					},
					"position": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string"
					},
					"lastModified": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"deckId",
					"wordOriginal",
					"wordTranslation",
					"position",
					"createdAt",
					"lastModified"
				],
				"type": "object",
				"additionalProperties": true
			},
			"DeckSummaryOptionalReturn": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"languageFrom": {
						"type": "string"
					},
					"languageTo": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"lastModified": {
						"type": "string"
					},
					"wordpairCount": {
						"type": "number",
						"format": "double"
					},
					"wordpairs": {
						"items": {
							"$ref": "#/components/schemas/WordPairEntity"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"languageFrom",
					"languageTo",
					"createdAt",
					"lastModified",
					"wordpairCount"
				],
				"type": "object",
				"additionalProperties": true
			},
			"apiSuccessResponse_DeckSummaryOptionalReturn_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/DeckSummaryOptionalReturn"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"DeckOptionalReturn": {
				"description": "DeckOptionalReturn is used by deck insertion and update endpoints to return their\nintended result. It is similar to DeckDetail, but the last two fields (wordpairCount and wordpairs)\nare optional. This offers flexibility, for instance, when the full list of wordpairs isn't returned.",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"languageFrom": {
						"type": "string"
					},
					"languageTo": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"lastModified": {
						"type": "string"
					},
					"wordpairCount": {
						"type": "number",
						"format": "double"
					},
					"wordpairs": {
						"items": {
							"$ref": "#/components/schemas/WordPairEntity"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"languageFrom",
					"languageTo",
					"createdAt",
					"lastModified"
				],
				"type": "object",
				"additionalProperties": true
			},
			"apiSuccessResponse_DeckOptionalReturn_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/DeckOptionalReturn"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"DeckCreateInput": {
				"properties": {
					"name": {
						"type": "string"
					},
					"languageFrom": {
						"type": "string"
					},
					"languageTo": {
						"type": "string"
					},
					"wordpairs": {
						"items": {
							"$ref": "#/components/schemas/WordPairInput"
						},
						"type": "array",
						"description": "Optional array of wordpairs to be inserted along with the deck."
					}
				},
				"required": [
					"name",
					"languageFrom",
					"languageTo"
				],
				"type": "object",
				"additionalProperties": true
			},
			"WordPairUpdateInput": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"deckId": {
						"type": "number",
						"format": "double"
					},
					"wordOriginal": {
						"type": "string"
					},
					"wordTranslation": {
						"type": "string"
					},
					"position": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"wordOriginal",
					"wordTranslation"
				],
				"type": "object",
				"additionalProperties": true
			},
			"DeckUpdateInput": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"languageFrom": {
						"type": "string"
					},
					"languageTo": {
						"type": "string"
					},
					"wordpairs": {
						"items": {
							"$ref": "#/components/schemas/WordPairUpdateInput"
						},
						"type": "array",
						"description": "Optional array of wordpairs. If not provided, only deck metadata is updated."
					}
				},
				"required": [
					"id",
					"name",
					"languageFrom",
					"languageTo"
				],
				"type": "object",
				"additionalProperties": true
			},
			"apiSuccessResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"apiSuccessResponse_WordPairEntity-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/WordPairEntity"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/decks/generate": {
			"post": {
				"operationId": "GenerateDeck",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/apiSuccessResponse_LLMDeck_"
								}
							}
						}
					}
				},
				"tags": [
					"Deck"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GenerateDeckRequest"
							}
						}
					}
				}
			}
		},
		"/decks/refine": {
			"post": {
				"operationId": "RefineDeck",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/apiSuccessResponse_LLMDeck_"
								}
							}
						}
					}
				},
				"tags": [
					"Deck"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RefineDeckRequest"
							}
						}
					}
				}
			}
		},
		"/decks/extract-name": {
			"post": {
				"operationId": "ExtractName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/apiSuccessResponse_ExtractNameResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Deck"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ExtractNameRequest"
							}
						}
					}
				}
			}
		},
		"/decks": {
			"get": {
				"operationId": "GetAllDecks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/apiSuccessResponse_DeckSummary-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Deck"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateDeck",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/apiSuccessResponse_DeckOptionalReturn_"
								}
							}
						}
					}
				},
				"tags": [
					"Deck"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeckCreateInput"
							}
						}
					}
				}
			}
		},
		"/decks/{deckId}": {
			"get": {
				"operationId": "GetDeckById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/apiSuccessResponse_DeckSummaryOptionalReturn_"
								}
							}
						}
					}
				},
				"description": "Get a deck by id.\nIf \"includeWordpairs\" is set to true, fetch the deck details with its associated wordpairs.",
				"tags": [
					"Deck"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deckId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "includeWordpairs",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateDeck",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/apiSuccessResponse_DeckOptionalReturn_"
								}
							}
						}
					}
				},
				"tags": [
					"Deck"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deckId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeckUpdateInput"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteDeck",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/apiSuccessResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Deck"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deckId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/decks/{deckId}/wordpairs": {
			"get": {
				"operationId": "GetDeckWordpairs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/apiSuccessResponse_WordPairEntity-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Deck"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deckId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateWordpairs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/apiSuccessResponse_WordPairEntity-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Deck"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deckId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/WordPairInput"
								},
								"type": "array"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateWordpairs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/apiSuccessResponse_WordPairEntity-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Deck"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deckId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/WordPairUpdateInput"
								},
								"type": "array"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteWordpairs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/apiSuccessResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Deck"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deckId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}